Create separate module for search family advice
