+++ src/pages/families/index.hbs
@@ -9,6 +9,7 @@ section: families
 page: families
 ---
 
+{{!-- TODO: Move styles to the separate css file --}}
 <style>
 #search-wrapper {
   margin-left: 100px;
@@ -87,12 +88,5 @@ page: families
   </div>
 </div>
 
-<div id="search-wrapper">
-  <div class="dropdown">
-    <input type="text" placeholder="Search our advice" class="form__input" id="search" data-bind="textInput: searchQuery, hasFocus: isSelectedSearchInput">
-    <div class="dropdown-content" data-bind="visible: showFilteredAdvice, foreach: filteredAdvice" >
-      <a href="#" data-bind="text: breadcrumbs, attr: { 'data-id': id }"></a>
-    </div>
-  </div>
-</div>
+{{> search-family-advice }}
 {{> loading-spinner}}
