+++ src/partials/search-family-advice.hbs
@@ -1,0 +1,8 @@
+<div id="search-wrapper" data-bind="with: searchFamilyAdvice">
+  <div class="dropdown">
+    <input type="text" placeholder="Search our advice" class="form__input" id="search" data-bind="textInput: searchQuery, hasFocus: isSelectedSearchInput">
+    <div class="dropdown-content" data-bind="visible: showFilteredAdvice, foreach: filteredAdvice" >
+      <a href="#" data-bind="text: breadcrumbs, attr: { 'data-id': id }"></a>
+    </div>
+  </div>
+</div>
